+++ src/Decomposition/TwinNumbers.java
@@ -1,7 +1,5 @@
 package Decomposition;
 
-import static Decomposition.ThreePrimeNumber.isPrime;
-
 public class TwinNumbers {
     public static int primeTwins(int amount) {
         int counter = counterOfPrimeNumbers(amount);
@@ -18,6 +16,14 @@ public class TwinNumbers {
             }
         }
         return counter;
+    }
+    public static boolean isPrime(int a) {
+        for (int i = 2; i < a; i++) {
+            if (a % i == 0) {
+                return false;
+            }
+        }
+        return true;
     }
 }
